perf(parser): extract the import resolver out of main()
main() carried 631 lines of import resolution: directory scanning, path
normalization, cycle and diamond detection, name mangling and AST
merging, plus E6001, E6005, E6011 and W2014. Another 156 lines of import
cache and directory scan helpers sat above it, used by nothing else.

Moved both into parser/imports.c behind imports_resolve(), which returns
the file-to-module attribution the type checker needs as a struct rather
than through six out-variables threaded down main(). read_file moved to
platform as gray_read_file since both callers now need it.

main() drops from 1244 lines to 606, main.c from 1706 to 854. No behavior
change: the old and new compilers produce byte-identical output across
all 1692 integration test files, 129 of them multi-file import tests.
